refactor: use a common components
