+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export { DataFormList } from './DataFormList/DataFormList';
 export { DataForm } from './DataForm/DataForm';
+export { FormInput } from './FormInput/FormInput';
+export { FormSelect } from './FormSelect/FormSelect';
