--- src/features/uncontrolledForm/components/FormInput.tsx
@@ -1,33 +1,0 @@
-import type { FC, PropsWithChildren } from 'react';
-import type { IFormInput } from '../../../types/forms';
-
-interface FormInputProps extends IFormInput {
-  error?: string;
-}
-
-export const FormInput: FC<PropsWithChildren<FormInputProps>> = ({
-  name,
-  placeholder,
-  title,
-  type,
-  error,
-  children,
-}) => {
-  return (
-    <div className="flex flex-col gap-1">
-      <label htmlFor={name} className="font-medium">
-        {title}
-      </label>
-      <input
-        id={name}
-        type={type}
-        name={name}
-        placeholder={placeholder}
-        autoComplete={'on'}
-        className={`border-2 rounded text-gray-700 border-gray-300 focus:border-blue-500 outline-0 px-2 py-1 transition-colors ${error && 'border-red-500'}`}
-      />
-      {children}
-      <p className="text-xs text-red-500 min-h-4">{error}</p>
-    </div>
-  );
-};
