+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { encodeImageFile } from './encodeImageFile';
+export { evaluatePasswordStrength } from './evaluatePasswordStrength';
